move "api helpers" into a separate file for readability
